Final Changes

+++ EmployeeTest/Models/EmployeeModel.cs
@@ -9,16 +9,16 @@ namespace EmployeeTest.Models
     public class EmployeeModel
     {
         [JsonProperty("id")]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         [JsonProperty("employee_name")]
         public string EmployeeName { get; set; }
 
         [JsonProperty("employee_salary")]
-        public long EmployeeSalary { get; set; }
+        public decimal EmployeeSalary { get; set; }
 
         [JsonProperty("employee_age")]
-        public long EmployeeAge { get; set; }
+        public int EmployeeAge { get; set; }
 
         [JsonProperty("profile_image")]
         public string ProfileImage { get; set; }
